MainMenu.unk_81fc must be int, otherwise assembly doesn't match

+++ src/th06_102h.exe/types.h
@@ -73087,7 +73087,7 @@ struct MainMenu {
     enum GameState gameState;
     int stateTimer;
     int idleFrames;
-    float unk_81fc;
+    int unk_81fc;
     D3DCOLOR maybe_menu_text_color;
     D3DCOLOR color2;
     D3DCOLOR color1;
